% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpca-package.R
\docType{package}
\name{morpca-package}
\alias{morpca-package}
\title{Manifold Optimization for Robust PCA}
\description{
Implements the manifold optimization procedure derived by
Teng Zhang and Yi Yang (2018) ... TODO...
}
\examples{
\dontrun{
TODO...
}

}
\author{
David Fleischer

Maintainer: David Fleischer <david.p.fleischer@gmail.com>
}
\keyword{package}
\keyword{pca}
