% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Percentile thresholding of matrix entries}
\usage{
threshold(X, gamma)
}
\arguments{
\item{X}{Input real-valued matrix.}

\item{gamma}{Thresholding percentile (in \eqn{[0, 1]}).}
}
\value{
Returns a real-valued matrix with \eqn{\gamma}-thresholded
entries.
}
\description{
Hard-thresholds the entries of a given real-valued matrix \eqn{X}
according to some thresholding percentile \eqn{\gamma} for which
the \eqn{(i, j)}-th entry of of \eqn{X} is set to zero if it
simultaneously exceeds the \eqn{(1-\gamma)}-th percentile of the
absolute values in row \eqn{i} and column \eqn{j}. That is,
\eqn{X_{ij}} is set to zero if it is among the most extreme
\eqn{\gamma\times 100}% of values within its row and column.
}
\details{
The hard-thresholding procedure will set the \eqn{(i,j)}-th
entry of X to 0 if
\deqn{|X_{i,j}| > |X^{[\gamma]}_{i,.}|}
and
\deqn{|X_{i,j}| > |X^{[\gamma]}_{.,j}|,}
where \eqn{X_{i,.}} denotes the \eqn{i}-th row of \eqn{X}, and
\eqn{X_{.,j}} represents the \eqn{j}-th column of \eqn{X}. Additionally,
\eqn{X^{[\gamma]}_{i,.}} and \eqn{X^{[\gamma]}_{.,j}} represent the
\eqn{\gamma}-th percentile of the absolute values of the entries of the
\eqn{i}-th row and the \eqn{j}-th column, respectively.
}
