% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Hard-Thresholding}
\usage{
threshold(X, gamma, sparsity)
}
\arguments{
\item{X}{Input real-valued data matrix.}

\item{gamma}{Thresholding percentile.}

\item{sparsity}{TO DO...}
}
\value{
Returns a real-valued matrix with \eqn{\gamma}-thresholded
entries.
}
\description{
Computes the hard-thresholding of a given real-valued matrix \eqn{X}
according to thresholding percentile \eqn{\gamma}.
}
\details{
The hard-thresholding procedure will set the \eqn{(i,j)}-th
entry of X to 0 if
\deqn{|X_{i,j}| > |X^{[\gamma]}_{i,.}|}
and
\deqn{|X_{i,j}| > |X^{[\gamma]}_{.,j}|,}
where \eqn{X_{i,.}} denotes the \eqn{i}-th row of \eqn{X}, and
\eqn{X_{.,j}} represents the \eqn{j}-th column of \eqn{X}. Additionally,
\eqn{X^{[\gamma]}_{i,.}} and \eqn{X^{[\gamma]}_{.,j}} represent the
\eqn{\gamma}-th percentile of the absolute values of the entries of the
\eqn{i}-th row and the \eqn{j}-th column, respectively.

That is, the elements of \eqn{X} which simultaneously exceed the
\eqn{\gamma}-th most extreme fraction of entries (in terms of absolute
value) in the corresponding rows and columns of \eqn{X} will be removed
and replaced with 0.
}
